move getTableConfiguration to helpers
